requirements.txt
fastapi
uvicorn
PyPDF2
openai
groq
sentence-transformers
chromadb
python-multipart
pandas
openpyxl
streamlit
requests
python-dotenv
